convert selects assign names tests to vitest
